requirements/tests.txt
molecule @ git+https://github.com/ansible-community/molecule@v24.8.0
molecule-podman==2.0.3
pytest-testinfra==10.1.1
